Refactor: Restructure server into feature folders; add sqlx postgres dependency
Move transactions HTTP handlers into src/features/transactions so
main.rs stays a thin bootstrap, and each feature registers its own
routes via a configure(cfg) function.
