+++ server/src/features/transactions/mod.rs
@@ -1,0 +1,3 @@
+mod handlers;
+
+pub use handlers::configure;
